Add files via upload
These are the main python files.

flask_app.py is the application itself.
AIDSAnalysisProcedures and AIDSAnalysisProcedures2 are packages for code to automatically generate most of the plots.  The main difference between them is AIDSAnalysisProcedures is designed for the first city chosen by the user and AIDSAnalysisProcedures2 is designed for the second city chosen.  AIDSAnalysisProcedures also has one extra function used by flask_app.py to create meshes for the contour plots.